research(loop-interface): name loop semantics; carve out core/ and service/
Split TurnDependencies out of LoopBackend so the intra-turn step function
next(dependencies, input) can only generate, execute tools, spawn children,
and write stream events; checkpoint, receive, finish, and spawnTurn stay
with the loop drivers. runTurn now drives named phases — initiate, advance,
settle — and the step result reuses the iterator protocol's done shape.

The step body is written as a minimal harness loop (pi's shape): generate;
a finish completes the turn; an approval parks it; otherwise every subagent
spawns before the first tool executes and child results fold back in request
order. Eager spawning is now a preserved semantic, not an implementation
accident. Stream.append becomes Stream.write; generationOrdinal becomes
stepOrdinal (operation-ID values unchanged).

Directories now mirror the ownership story: core/ holds the programs, the
step function, and the executable contract; service/ holds the scripted
effects and event store; checkpoint protocol, wire codec, and conformance
stay at the root between core and the adapters. The research doc gains a
named-loop-semantics section grounding each term in the harness it comes
from (AI SDK step, iterator next/done, injected dependencies) and updated
file links.
