+++ packages/eve/src/internal/testing/loop-prototype/core/index.ts
@@ -1,2 +1,3 @@
 export { runSession } from "./session-program.js";
 export { runTurn } from "./turn-program.js";
+export { next } from "./turn-step.js";
