+++ packages/eve/src/internal/testing/loop-prototype/core/program-effects.ts
@@ -3,28 +3,28 @@ import type {
   ApprovalRequest,
   GenerateInput,
   GeneratedTurn,
-  LoopBackend,
   OperationId,
   RequestResult,
   ToolRequest,
+  TurnDependencies,
 } from "./types.js";
 
 export async function generate(
-  backend: LoopBackend,
+  dependencies: TurnDependencies,
   input: GenerateInput,
 ): Promise<{ readonly operationId: OperationId; readonly output: GeneratedTurn }> {
   return {
     operationId: generateOperationId(input),
-    output: await backend.generate(input),
+    output: await dependencies.generate(input),
   };
 }
 
 export async function executeTool(
-  backend: LoopBackend,
+  dependencies: TurnDependencies,
   request: ApprovalRequest | ToolRequest,
 ): Promise<{ readonly operationId: OperationId; readonly output: RequestResult }> {
   return {
     operationId: executeToolOperationId(request),
-    output: await backend.executeTool(request),
+    output: await dependencies.executeTool(request),
   };
 }
