--- packages/eve/src/internal/testing/loop-prototype/service.ts
@@ -1,4 +1,0 @@
-export type { PrototypeService } from "./service-contract.js";
-export { DeclaredEffectFailure, EffectProtocolError } from "./service-effects.js";
-export { MemoryPrototypeService } from "./service-memory.js";
-export { SqlitePrototypeService } from "./service-sqlite.js";
