+++ packages/eve/src/internal/testing/loop-prototype/service/index.ts
@@ -1,0 +1,4 @@
+export type { PrototypeService } from "./contract.js";
+export { DeclaredEffectFailure, EffectProtocolError } from "./effects.js";
+export { MemoryPrototypeService } from "./memory.js";
+export { SqlitePrototypeService } from "./sqlite.js";
